fix: Fixes HtmlLocalized color for gump layouts

+++ Projects/Server/Gumps/StaticGumpBuilder.cs
@@ -227,7 +227,7 @@ public ref struct StaticGumpBuilder
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void AddHtmlLocalized(
-        int x, int y, int width, int height, int number, short color, bool background = false, bool scrollbar = false
+        int x, int y, int width, int height, int number, int color, bool background = false, bool scrollbar = false
     ) => _gumpBuilder.AddHtmlLocalized(x, y, width, height, number, color, background, scrollbar);
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
